+++ src/podofo/main/PdfPainter.h
@@ -200,6 +200,8 @@ private:
  *
  * All functions that take coordinates expect these to be in PDF User Units. Keep in mind that PDF has
  * its coordinate system origin at the bottom left corner.
+ * \remarks It should not be generally needede but you can cast the instance to
+ * PdfContentStreamOperators to access low level PDF operators
  */
 class PODOFO_API PdfPainter final : public PdfContentStreamOperators
 {